+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## [1.3.15] - 2020-11-04
+
+### Added
+
+- Inline component in @react-corn/material-ui
+- Make Autocomplete inlineable
+
 ## [1.3.14] - 2020-08-27
 
 ### Fixed
@@ -160,7 +167,8 @@ Bad release
 
 ## [1.0.1] - 2020-06-04
 
-[unreleased]: https://github.com/paradoxxxzero/react-corn/compare/v1.3.14...HEAD
+[unreleased]: https://github.com/paradoxxxzero/react-corn/compare/v1.3.15...HEAD
+[1.3.15]: https://github.com/paradoxxxzero/react-corn/compare/v1.3.14...1.3.15
 [1.3.14]: https://github.com/paradoxxxzero/react-corn/compare/v1.3.13...1.3.14
 [1.3.13]: https://github.com/paradoxxxzero/react-corn/compare/v1.3.12...1.3.13
 [1.3.12]: https://github.com/paradoxxxzero/react-corn/compare/v1.3.11...1.3.12